implement compound hets under new system
